requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.6.2
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
distlib==0.3.8
dnspython==2.6.1
email-validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.4
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
identify==2.5.36
idna==3.7
iniconfig==2.0.0
isort==5.13.2
jinja2==3.1.4
markdown-it-py==3.0.0
markupsafe==2.1.5
mdurl==0.1.2
nodeenv==1.9.1
orjson==3.10.5
packaging==24.1
pillow==10.3.0
pip==24.1.1
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.7.1
pydantic==2.7.4
pydantic-core==2.18.4
pydantic-settings==2.3.4
pygments==2.18.0
pytest==8.2.2
python-dotenv==1.0.1
python-multipart==0.0.9
pyyaml==6.0.1
requests==2.32.3
rich==13.7.1
ruff==0.5.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.37.2
typer==0.12.3
typing-extensions==4.12.2
ujson==5.10.0
urllib3==2.2.2
uv==0.2.18
uvicorn==0.30.1
uvloop==0.19.0
virtualenv==20.26.3
watchfiles==0.22.0
websockets==12.0
